feat: implement user authorization and protect routes with middleware

+++ controllers/taskController.js
@@ -1,9 +1,11 @@
+const User = require('../models/userModel')
+
 const asyncHandler = require('express-async-handler');
 
 const Task = require('../models/taskModel');
 
 const getTasks = asyncHandler(async (req, res)=>{
-    const tasks = await Task.find();
+    const tasks = await Task.find({user: req.user._id});
     res.status(200).json(tasks);
 })
 
@@ -15,7 +17,7 @@ const setTask = asyncHandler(async (req, res)=>{
     }
 
     const task = await Task.create({
-        text: req.body.text
+        text: req.body.text, user:req.user._id
     })
     // console.log(req.body);
     res.status(200).json(task)
@@ -27,6 +29,17 @@ const updateTask = asyncHandler(async (req, res)=>{
         res.status(400)
         throw new Error('Task not found')
     }
+
+    const user = await User.findById(req.user._id)
+    if(!user){
+        res.status(401)
+        throw new Error('No such user found')
+    }
+    if(task.user.toString()!== user.id){
+        res.status(401)
+        throw new Error('User is not authorized to update')
+    }
+
     const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json(updatedTask)
 })
@@ -37,6 +50,18 @@ const deleteTask = asyncHandler(async (req, res)=>{
         res.status(400)
         throw new Error('Task not found')
     }
+
+    const user = await User.findById(req.user._id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('No such user found')
+    }
+    if(task.user.toString()!== user._id.toString()){
+        res.status(401)
+        throw new Error('User is not authorized to delete')
+    }
+
     await Task.findByIdAndDelete(req.params.id)             
     res.status(200).json({message: `Task ${req.params.id} deleted.`})
 })
